organize commponents
